New helper scripts
Calculate Calibrated Temperatures - a GUI to input right and left arm calibrated values and calculate the range for the psychometric function in the main experiment.

Load Thermode - loads thermode functions and gives example stimulation. Helps to check the thermodes are working and loaded in the correct order, without having to run the main experiment script.